Remove AutoPing responsibility from the user object
The User object already knows a lot (too much) about the Manager as it
is setting itself as the current user but starting Auto Ping should be
the responsibility of the Manager itself.

In the Current User setter, the Manager will start Auto Ping if the
assigned user is logged in. This means that if you pass a user which is
not signed in, it will never Auto Ping (because the Manager is not
looking for changes on the IsAuthenticated property. However, this
scenario will only happen if the current user is assigned by hand (the
API itself will only pass an already logged in user). Thus, if someone
wants to mess with the API, it should be there responsibility to handle
such thing.