Expose return value for dismissible windows
Whether the windows dismiss callback should return success or failure
depend on the window type itself.
For example, dismissing the login window is a failure as the login was
unsuccessful but dismissing the trophy window is a success as the window
was shown properly.

Note: The UI close button should return the same value.