Short-circuit leaderboard window show animation if scores are received
Set animation to jump to `show` from any state but `dismiss` if the score are received.
It solves a race condition where the scores would be fetched quickly (especially when using caching)
and the window would still be appearing and wouldn't listen to the `unlock` (i.e. show) event.

Fix #47 and Fix #65